folder and file upload torrents feature

+++ peercds/src/main/webapp/js/peer.js
@@ -15,11 +15,13 @@
 		currentFile:"",
 		currentTorrentState:''
 	};
-	
+	console.log('current file is '+myobject.currentFile);
+	if(myobject.currentFile==""){}
 	$('#torrent-info').hide();
 	$('#page-wrapper').click(function(e){
 		$('#torrent-info').hide();
 		$('#torrent-rows > tr').removeClass('info');
+		$('.dropdown-menu').hide();
 		myobject.selectedFile="";
 		e.stopPropagation();
 	});
@@ -30,11 +32,16 @@
 		myobject.selectedFile="";
 		e.stopPropagation();
 	});
+	
+	$('document').click(function (e) {
+		$('.dropdown-menu').hide();
+	});
 
 $(function(){
 	timer();
 	setInterval(timer, 3000);
 	$("#torrent_table").on('click','#torrent-rows > tr',function(e){
+		console.log('current file is '+myobject.currentFile);
 		console.log(myobject.currentFile[0].uuid);
 		var status=myobject.currentFile[0].status;
 		
@@ -104,13 +111,8 @@ $(function(){
     	timer();
     });
 
-	$('#share-torrent').click(function(e){
-		$('#upload_torrent').modal('show');
-		$in=$(this);
-		$in.val("");
-		
+	
 		$('#create_torrent').change(function(e){
-			$('#upload_torrent').modal('hide');
 			$in=$(this);
 			console.log($in.val());
 			$('#tracker_url_modal').modal('show');
@@ -125,27 +127,34 @@ $(function(){
 	        	});
 	    	}
 			$in.val("");
+			
 			e.stopPropagation();
 		});
-		
+    
+	
+	
 		$('#file_input').change(function(e){
 			$('#upload_torrent').modal('hide'); 
-			console.log($('#dir-tree').find('li:first-child > a').html());
 			var folder_name=$('#dir-tree').find('li:first-child > a').html();
+			alert(folder_name);
 			$('#tracker_url_modal').modal('show');
-			if(folder_name.length>0){
-	    		$('#url-submit').click(function(){
-	        		var trackerurl=$('#tracker_url').val();
-	    	       	 if(trackerurl!==""){
-	    	       		 peerApi.createTorrent(folder_name,trackerurl);
-	    	       	 }
-	        	});
-	    	}
+			$('#url-submit').click(function(){
+	    		var trackerurl=$('#tracker_url').val();
+		       	 if(trackerurl!==""){
+		       		 peerApi.createTorrent(folder_name,trackerurl);
+		       	 }
+	    	});
+			$('#dir-tree').empty();
 			e.stopPropagation();
 		});
-
-    	e.stopPropagation();
-    });
+	
+	
+	
+	
+	
+	
+	
+	
 
     $('input[type=file]#download_torrent').change(function(e){
     	$in=$(this);    	
@@ -172,11 +181,13 @@ $(function(){
     	}
     });
     
-    $('#Delete').click(function(){
+    $('#Delete').click(function(e){
     	if(myobject.currentFile!==""){
-    		console.log('file is selected');
-    		peerApi.deleteTorrent(peerApi.uuid);
+    		if(confirm("Please confirm to delete the file")==true){
+        		peerApi.deleteTorrent(peerApi.uuid);
+    		}
     	}
+    	e.stopPropagation();
     });
         
 });
